requirements.txt
flask==2.3.3
pyyaml==6.0.1
requests==2.31.0
llama-cpp-python==0.2.19; python_version>="3.8"
watchdog==3.0.0
Werkzeug==2.3.7
Jinja2==3.1.3
itsdangerous==2.1.2
MarkupSafe==2.1.5
click==8.1.7
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.4
urllib3==2.0.7